+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog
 * Fixed type check for ``water_template``.
 * Add support for simulations in the osmotic ensemble.
 * Fixed non-uniform bulk insertion positions caused by use of normal rather than uniform random numbers.
+* Add methods to update the system with the current water state and return system without ghost waters.
 
 [2025.2.0](https://github.com/openbiosim/loch/compare/2025.1.0...2025.2.0) - Feb 2026
 -------------------------------------------------------------------------------------